implemented creating remote rooms refactor
