+++ src/Empire/ProcessFlagHelpers/ProcessDefaultRemoteRoom.ts
@@ -1,4 +1,4 @@
-import { UserException, EmpireHelper, MemoryApi_Room, MemoryApi_Empire, EmpireApi } from "Utils/Imports/internals";
+import { UserException, EmpireHelper, MemoryApi_Room, MemoryApi_Empire, EmpireApi, REMOTE_ENERGY } from "Utils/Imports/internals";
 import _ from "lodash";
 
 export class ProcessDefaultRemoteRoom implements IFlagProcesser {
@@ -15,70 +15,16 @@ export class ProcessDefaultRemoteRoom implements IFlagProcesser {
      * @param flag
      */
     public processFlag(flag: Flag): void {
-        // Get the host room and set the flags memory
-        const dependentRoom: Room = Game.rooms[EmpireHelper.findDependentRoom(flag.pos.roomName)];
-        const flagTypeConst: FlagTypeConstant | undefined = EmpireHelper.getFlagType(flag);
-        const roomName: string = flag.pos.roomName;
-        Memory.flags[flag.name].complete = false;
-        Memory.flags[flag.name].processed = true;
-        Memory.flags[flag.name].timePlaced = Game.time;
-        Memory.flags[flag.name].flagType = flagTypeConst;
-        Memory.flags[flag.name].flagName = flag.name;
-
         switch (flag.secondaryColor) {
             // Create an energy focused remote room
             case COLOR_YELLOW:
-                EmpireApi.createRemoteRoomInstance(roomName);
+                EmpireApi.createRemoteRoomInstance(flag, REMOTE_ENERGY);
                 break;
 
             // Clear an existing remote room
             case COLOR_RED:
-                EmpireApi.removeRemoteRoom(roomName);
+                EmpireApi.removeRemoteRoomInstance(flag);
                 break;
         }
-
-        // Create the RemoteFlagMemory object for this flag
-        const remoteFlagMemory: RemoteFlagMemory = {
-            flagName: flag.name,
-            flagType: flagTypeConst
-        };
-
-        // If the dependent room already has this room covered, set the flag to be deleted and throw a warning
-        const existingDepedentRemoteRoomMem: RemoteRoomMemory | undefined = _.find(
-            MemoryApi_Room.getRemoteRooms(dependentRoom),
-            (rr: RemoteRoomMemory) => {
-                if (rr) {
-                    return rr.roomName === roomName;
-                }
-                return false;
-            }
-        );
-
-        if (existingDepedentRemoteRoomMem) {
-            Memory.flags[flag.name].complete = true;
-            throw new UserException(
-                "Already working this dependent room!",
-                "The room you placed the remote flag in is already being worked by " +
-                existingDepedentRemoteRoomMem.roomName,
-                ERROR_WARN
-            );
-        }
-
-        // Otherwise, add a brand new memory structure onto it
-        const remoteRoomMemory: RemoteRoomMemory = {
-            sources: { cache: Game.time, data: 1 },
-            hostiles: { cache: Game.time, data: null },
-            structures: { cache: Game.time, data: null },
-            roomName: flag.pos.roomName,
-            flags: [remoteFlagMemory],
-            reserveTTL: 0,
-            reserveUsername: undefined
-        };
-
-        MemoryApi_Empire.createEmpireAlertNode(
-            "Remote Flag [" + flag.name + "] processed. Host Room: [" + dependentRoom.name + "]",
-            10
-        );
-        dependentRoom.memory.remoteRooms!.push(remoteRoomMemory);
     }
 }
